requirements.txt
pydantic
openai
google-generativeai
requests
sqlalchemy
python-dotenv
tavily-python
